Add random number seed as optional user setting.
Mostly for debugging. If no seed is provided, use the system clock.

I noticed that:

srand(time(NULL));

was used multiple times in both:

main.cpp
siman_cal_par.cpp

so that analyses could not be replicated. Since srand is called in main, it doesn't need to be called again anywhere else. I've therefore commented-out all but the first occurrence.